# lib/api-zod/src/generated/types/identityVerificationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Companion REST API
 * OpenAPI spec version: 0.1.0
 */
import type { IdentityVerificationInputRole } from './identityVerificationInputRole';

export interface IdentityVerificationInput {
  role: IdentityVerificationInputRole;
  /** @minLength 1 */
  identifier: string;
  /** @minLength 7 */
  mobile: string;
}