# lib/api-zod/src/generated/types/identityVerificationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Companion REST API
 * OpenAPI spec version: 0.1.0
 */

export interface IdentityVerificationResponse {
  maskedMobile: string;
  resendAvailableIn: number;
}