# lib/api-zod/src/generated/types/otpVerificationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Companion REST API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpVerificationInput {
  /**
     * @minLength 6
     * @maxLength 6
     */
  otp: string;
}